Made the function of displaying a random name for users as their username

+++ src/SoftwareBookList/Models/Rating.cs
@@ -9,8 +9,9 @@ namespace SoftwareBookList.Models
 		public int RatingID { get; set; }
 
 		[Required]
-		[Range(1, 10, ErrorMessage = "RatingValue must be between 1 and 10.")]
-		public int RatingValue { get; set; }
+		[Range(1.0, 10.0, ErrorMessage = "RatingValue must be between 1 and 10.")]
+		// This makes it so, the rating can be a whole or decimal value.
+		public double RatingValue { get; set; }
 
 
 		// Navigation property for the one-to-many relationship with Review
